+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@
   * Added `ConvGeneralOp` to infer the output shape of a convolution.
 * Package `backends/simplego`:
   * Implemented `ConvGeneralDilated` operation.
+* Package `types/shapes`: 
+  * `Shape.Iter()` and `Shape.IterOn()` also yields the flat index being iterated.
 
 # v0.21.1: 2025/08/16 Added Zero-dim tensors support and other small improvements.